pyproject.toml
[project]
name = "database-setup-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "duckdb>=1.1.3",
    "fastapi[standard]==0.115.6",
    "jinja2==3.1.4",
    "psycopg2-binary==2.9.10",
    "ruff>=0.8.4",
]
